backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.0

# Database
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
asyncpg==0.30.0

# Config & validation
pydantic-settings==2.6.1

# Auth
python-jose[cryptography]==3.3.0
bcrypt==5.0.0

# Redis
redis[asyncio]==5.2.0

# HTTP client (WhatsApp API calls)
httpx==0.27.2

# Email
aioimaplib==1.1.0
aiosmtplib==3.0.1

# Parsing
beautifulsoup4==4.12.3

# Logging
structlog==24.4.0

# Rate limiting
slowapi==0.1.9
